cli/pyproject.toml
[project]
name = "cli"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "connectorx>=0.4.0",
    "polars>=1.16.0",
    "pyarrow>=18.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
